Improve the PathParameter tests Remove unusable PathParameter initializer
